Added booleans
