+++ contracts/wasm/tokenregistry/go/tokenregistry/lib.go
@@ -47,6 +47,7 @@ type TransferOwnershipContext struct {
 
 func funcTransferOwnershipThunk(ctx wasmlib.ScFuncContext) {
 	ctx.Log("tokenregistry.funcTransferOwnership")
+
 	// TODO the one who can transfer token ownership
 	ctx.Require(ctx.Caller() == ctx.ContractCreator(), "no permission")
 
@@ -70,6 +71,7 @@ type UpdateMetadataContext struct {
 
 func funcUpdateMetadataThunk(ctx wasmlib.ScFuncContext) {
 	ctx.Log("tokenregistry.funcUpdateMetadata")
+
 	// TODO the one who can change the token info
 	ctx.Require(ctx.Caller() == ctx.ContractCreator(), "no permission")
